pyproject.toml
[project]
name = "toshokan"
version = "0.1.0"
description = "App for assiting in learning the Japanese language"
authors = [
    {name = "ishish.io",email = "ts@ishish.io"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio (>=5.38.2,<6.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
